requirements.txt
python-telegram-bot[job-queue]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client
